+++ java-manta-it/src/test/java/com/joyent/manta/client/MantaClientDirectoriesIT.java
@@ -59,7 +59,7 @@ public class MantaClientDirectoriesIT {
     public void afterClass() throws IOException, MantaCryptoException {
         if (mantaClient != null) {
             mantaClient.deleteRecursive(testPathPrefix);
-            mantaClient.closeQuietly();
+            mantaClient.closeWithWarning();
         }
     }